More bioinf work

+++ docs/data/learn/Bioinformatics/input/ch4_code/src/NaiveSpectrumConvolution.py
@@ -5,7 +5,8 @@ from PrefixSumTheoreticalSpectrum import theoretical_spectrum, PeptideType
 from helpers.AminoAcidUtils import get_amino_acid_to_mass_table
 
 
-def spectrum_convolution(experimental_spectrum: List[int], min_mass=57, max_mass=200) -> List[int]:
+# MARKDOWN
+def spectrum_convolution(experimental_spectrum: List[float], min_mass=57.0, max_mass=200.0) -> List[float]:
     # it's expected that experimental_spectrum is sorted smallest to largest
     diffs = []
     for row_idx, row_mass in enumerate(experimental_spectrum):
@@ -15,6 +16,7 @@ def spectrum_convolution(experimental_spectrum: List[int], min_mass=57, max_mass
                 diffs.append(mass_diff)
     diffs.sort()
     return diffs
+# MARKDOWN
 
 
 if __name__ == '__main__':
